Added fiex spawnpoints in obstacleGenerator
You can now drop down PickUpCoords object to manipulate spawnPoints of obstacles. PickUpCoords is now set as a serialized field in obstacleGenerator script, but change only position of its children game object in order to change the spawn points. To modify the object that is being spawned, modify 'prefeb' field.